Break code into files
- bot and provider moved to services folder
- commands moved to its separate folder
- related code moved to its related file in the utils folder
